Add latest five moto and car posts endpoints
Introduced two new methods to the PostRepository: 'lastMoto' and 'lastCar', which return the five latest posts of type 'moto' and 'car' respectively. These methods are used in the newly created endpoints '/lastMoto' and '/lastCar' in the PostController. Also enhanced the destroy method in the PostController to provide more informative responses.