Update PostRepository query organization and filter criteria
Modified the PostRepository methods 'lastMoto' and 'lastCar' to utilize direct filtering of vehicle types. This was achieved by distinguishing between 'moto' or 'car' via the 'is_two_wheeled' boolean attribute, bolstering code readability and simplification. The changes also include an update to the ordering of fetched posts to be dependent on 'created_at' in descending order.