Replace ArticleResource with PostResource in PostController
This commits change the methods that previous used ArticleResource to use the new PostResource instead, specifically for the "lastMoto" and "lastCar" methods. This change is part of our effort to streamline our resources and ensure similar posts are using the same resource class, making our application more maintainable and scalable.